Fixed tests, removed visualization
