+++ test/data/test_poland.py
@@ -1,8 +1,7 @@
 from unittest import TestCase
 import pandas as pd
-import os
 from pathlib import Path
-from src.data.datasets import generations_configuration_xlsx
+from src.data.datasets import voivodship_cities_generations_configuration_xlsx
 from src.data import preprocessing_poland as poland
 
 
@@ -10,12 +9,12 @@ class TestGenerationConfiguration(TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         project_dir = Path(__file__).resolve().parents[2]
-        data_folder = project_dir / 'data' / 'processed' / 'poland' / 'DW'
+        data_folder = project_dir / 'data' / 'processed' / 'poland' / 'WW'
         assert data_folder.is_dir()
-        assert (data_folder / generations_configuration_xlsx.file_name).is_file()
+        assert (data_folder / voivodship_cities_generations_configuration_xlsx.file_name).is_file()
         cls.generations_configuration_df = pd.read_excel(
-            str(data_folder / generations_configuration_xlsx.file_name),
-            sheet_name=generations_configuration_xlsx.sheet_name)
+            str(data_folder / voivodship_cities_generations_configuration_xlsx.file_name),
+            sheet_name=voivodship_cities_generations_configuration_xlsx.sheet_name)
 
     def test_find_configurations_for_family_type_1_selfcontained(self):
         # df, headcount, family_type, relationship, house_master):
