+++ test/features/test_population_generator_common.py
@@ -1,11 +1,12 @@
 from unittest import TestCase
-from src.features import population_generator_common as population
-from src.features import entities
-from src.data import datasets
-import scipy.stats
+from pathlib import Path
+
 import numpy as np
 import pandas as pd
-from pathlib import Path
+import scipy.stats
+
+from src.features import entities
+from src.features import population_generator_common as population
 
 
 class TestNodesToDataFrame(TestCase):
@@ -26,21 +27,24 @@ class TestNodesToDataFrame(TestCase):
                               social_competence=0.3,
                               public_transport_duration=0,
                               public_transport_usage=False,
-                              household=1, profession=entities.PROFESSION_NOT_ASSIGNED)
+                              household=1, profession=entities.PROFESSION_NOT_ASSIGNED,
+                              age_generation=entities.AgeGroup.young.value)
         node2 = entities.Node(age=37, gender=entities.Gender.FEMALE,
                               employment_status=entities.EmploymentStatus.EMPLOYED,
                               social_competence=0.7,
                               public_transport_duration=30,
                               public_transport_usage=True,
-                              household=1, profession=7)
+                              household=1, profession=7,
+                              age_generation=entities.AgeGroup.middle.value)
         nodes = [node1, node2]
 
         df = population.nodes_to_dataframe(nodes)
 
         self.assertIsInstance(df, pd.DataFrame)
-        self.assertEqual(len(entities.columns), len(df.columns))
+        self.assertEqual(len(entities.columns) + 1, len(df.columns))  # age_generation is not in a final dataframe
         self.assertEqual(2, len(df.index))
-        self.assertEqual(entities.columns, df.columns.tolist())
+        for col in entities.columns:
+            self.assertIn(col, df.columns.tolist())
 
         self.assertNodeAndRowEqual(node1, df.iloc[0])
         self.assertNodeAndRowEqual(node2, df.iloc[1])
@@ -54,3 +58,45 @@ class TestNodesToDataFrame(TestCase):
         self.assertEqual(node.public_transport_duration, row[entities.prop_public_transport_duration])
         self.assertEqual(node.household, row[entities.prop_household])
         self.assertEqual(node.profession, row[entities.prop_profession])
+
+
+class TestDistributionRetrieval(TestCase):
+
+    def test_should_raise_error_on_unknown_datatype(self):
+        try:
+            population.get_distribution(1)
+            self.fail('Should have raised ValueError')
+        except ValueError as e:
+            self.assertTrue(str(e).startswith('Expected the name of a distribution'))
+
+    def test_should_raise_error_on_unknown_distribution(self):
+        try:
+            population.get_distribution('giberish')
+            self.fail('Should have raised an error')
+        except Exception as e:
+            self.assertEqual('module \'scipy.stats\' has no attribute \'giberish\'', str(e))
+
+    def test_should_return_normal_distribution(self):
+
+        distribution = population.get_distribution('norm')
+        self.assertIsInstance(distribution, scipy.stats._continuous_distns.norm_gen)
+
+
+class TestDrawingFromDistribution(TestCase):
+
+    def test_should_draw_10_samples_from_normal_distribution(self):
+        sample = population.sample_from_distribution(10, 'norm', 0, 1)
+        self.assertEqual(10, len(sample))
+        self.assertIsInstance(sample, np.ndarray)
+
+
+class TestCleanup(TestCase):
+
+    def test_age_transformation(self):
+        df = pd.read_excel(str(Path(__file__).resolve().parents[0] / 'test_cleanup_age_parsing.xlsx'))
+        result = population.age_range_to_age(df)
+
+        self.assertEqual(int, result[entities.prop_age].dtype)
+        for idx, row in result.iterrows():
+            self.assertGreaterEqual(row.age, row.age_expected)
+            self.assertLess(row.age, row.age_expected + 5)
