+++ test/features/test_population_generator_for_wroclaw.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from src.features import population_generator_for_wroclaw as population
+from src.features import population_generator_for_cities as population
 from src.features import entities
 from src.data import datasets
 import scipy.stats
@@ -7,35 +7,6 @@ import numpy as np
 import pandas as pd
 from pathlib import Path
 
-
-class TestDistributionRetrieval(TestCase):
-
-    def test_should_raise_error_on_unknown_datatype(self):
-        try:
-            population.get_distribution(1)
-            self.fail('Should have raised ValueError')
-        except ValueError as e:
-            self.assertTrue(str(e).startswith('Expected the name of a distribution'))
-
-    def test_should_raise_error_on_unknown_distribution(self):
-        try:
-            population.get_distribution('giberish')
-            self.fail('Should have raised an error')
-        except Exception as e:
-            self.assertEqual('module \'scipy.stats\' has no attribute \'giberish\'', str(e))
-
-    def test_should_return_normal_distribution(self):
-
-        distribution = population.get_distribution('norm')
-        self.assertIsInstance(distribution, scipy.stats._continuous_distns.norm_gen)
-
-
-class TestDrawingFromDistribution(TestCase):
-
-    def test_should_draw_10_samples_from_normal_distribution(self):
-        sample = population.sample_from_distribution(10, 'norm', 0, 1)
-        self.assertEqual(10, len(sample))
-        self.assertIsInstance(sample, np.ndarray)
 
 
 class TestSocialCompatenece(TestCase):
@@ -55,11 +26,11 @@ class TestNarrowHousemastersByAgeAndFamilyStructure(TestCase):
     def setUpClass(cls) -> None:
         test_file = Path(__file__).resolve().parents[0] / 'households_combinations.xlsx'
         project_dir = Path(__file__).resolve().parents[2]
-        voivodship_folder = project_dir / 'data' / 'processed' / 'poland' / 'D'
-        assert voivodship_folder.is_dir()
-        assert (voivodship_folder / datasets.households_by_master_xlsx.file_name).is_file()
-        cls.household_by_master = pd.read_excel(str(voivodship_folder / datasets.households_by_master_xlsx.file_name),
-                                                sheet_name=datasets.households_by_master_xlsx.sheet_name)
+        data_folder = project_dir / 'data' / 'processed' / 'poland' / 'DW'
+        assert (data_folder / datasets.voivodship_cities_households_by_master_xlsx.file_name).is_file()
+        cls.household_by_master = pd.read_excel(
+            str(data_folder / datasets.voivodship_cities_households_by_master_xlsx.file_name),
+            sheet_name=datasets.households_by_master_xlsx.sheet_name)
 
         assert test_file.is_file()
         cls.households_rows = pd.read_excel(str(test_file))
@@ -101,14 +72,4 @@ class TestPublicTransport(TestCase):
         self.assertEqual(len(duration.index), self.population_size)
 
 
-class TestCleanup(TestCase):
-
-    def test_age_transformation(self):
-        df = pd.read_excel(str(Path(__file__).resolve().parents[0] / 'test_cleanup_age_parsing.xlsx'))
-        result = population._age_range_to_age(df)
-
-        self.assertEqual(int, result[entities.prop_age].dtype)
-        for idx, row in result.iterrows():
-            self.assertGreaterEqual(row.age, row.age_expected)
-            self.assertLess(row.age, row.age_expected + 5)
